Fix 64-bit problem in recent patch.

+++ src/backend/access/gist/gistutil.c
@@ -8,7 +8,7 @@
  * Portions Copyright (c) 1994, Regents of the University of California
  *
  * IDENTIFICATION
- *			$PostgreSQL: pgsql/src/backend/access/gist/gistutil.c,v 1.27 2008/06/12 09:12:29 heikki Exp $
+ *			$PostgreSQL: pgsql/src/backend/access/gist/gistutil.c,v 1.28 2008/06/15 01:41:37 tgl Exp $
  *-------------------------------------------------------------------------
  */
 #include "postgres.h"
@@ -45,7 +45,7 @@ gistfillbuffer(Page page, IndexTuple *itup, int len, OffsetNumber off)
 		l = PageAddItem(page, (Item) itup[i], sz, off, false, false);
 		if (l == InvalidOffsetNumber)
 			elog(ERROR, "failed to add item to GiST index page, item %d out of %d, size %d bytes",
-				 i, len, sz);
+				 i, len, (int) sz);
 		off++;
 	}
 }
